@@ -30,25 +30,28 @@ import Foundation
  **Example**:
  ```swift
  let request: URLRequest = //
- let resource: Resource<String?> = Resource(request: request, parse: { data in
+ let resource: Resource<String?, NetworkError> = Resource(request: request, parse: { data in
     String(data: data, encoding: .utf8)
  })
  ```
  */
-public struct Resource<Model> {
+public struct Resource<Model, E: Error>: Sendable {
     /// The request to fetch the resource remote payload
     public let request: URLRequest
-    
+
     /// Parses data into given model.
-    public let parse: (_ data: Data) throws -> Model
-    
-    /// Creates a type safe resource, which can be used to fetch it with `NetworkService`
+    public let parse: @Sendable (_ data: Data) throws -> Model
+    public let mapError: @Sendable (_ networkError: NetworkError) -> E
+
+    /// Creates a type safe resource, which can be used to fetch it with NetworkService
     ///
     /// - Parameters:
-    ///   - request: The request to get the remote data payload
-    ///   - parse: Parses data fetched with the request into given Model
-    public init(request: URLRequest, parse: @escaping (Data) throws -> Model) {
+    /// - request: The request to get the remote data payload
+    /// - parse: Parses data fetched with the request into given Model
+
+    public init(request: URLRequest, parse: @escaping @Sendable  (Data) throws -> Model, mapError: @escaping @Sendable  (_ networkError: NetworkError) -> E) {
         self.request = request
         self.parse = parse
+        self.mapError = mapError
     }
 }
