@@ -26,265 +26,95 @@ import XCTest
 @testable import DBNetworkStack
 
 class NetworkServiceTest: XCTestCase {
-    
-    var networkService: NetworkService!
-    
-    var networkAccess = NetworkAccessMock()
-    
+
     let trainName = "ICE"
     
-    var resource: Resource<Train> {
+    var resource: Resource<Train, NetworkError> {
         let request = URLRequest(path: "train", baseURL: .defaultMock)
         return Resource(request: request, decoder: JSONDecoder())
     }
-    
-    override func setUp() {
-        networkService = BasicNetworkService(networkAccess: networkAccess)
-    }
-    
-    func testRequest_withValidResponse() {
-        //Given
-        networkAccess.changeMock(data: Train.validJSONData, response: .defaultMock, error: nil)
-        let expection = expectation(description: "loadValidRequest")
-        
-        //When
-        networkService.request(resource, onCompletionWithResponse: { train, response in
-            XCTAssertEqual(train.name, self.trainName)
-            XCTAssertEqual(response, .defaultMock)
-            expection.fulfill()
-            }, onError: { _ in
-                XCTFail("Should not call error block")
-        })
-        
-        waitForExpectations(timeout: 1, handler: nil)
-        
-        //Then
-        XCTAssertEqual(networkAccess.request?.url?.absoluteString, "https://bahn.de/train")
-    }
 
-    func testRequest_withNoDataResponse() {
+    
+    func testRequest_withValidResponse() async throws {
         //Given
-        networkAccess.changeMock(data: nil, response: nil, error: nil)
-        let expection = expectation(description: "testNoData")
-        
-        //When
-        var capturedError: NetworkError?
-        networkService.request(resource, onCompletion: { _ in
-            XCTFail("Should not call success block")
-            }, onError: { error in
-                capturedError = error
-                expection.fulfill()
-        })
-        
-        //Then
-        waitForExpectations(timeout: 1, handler: nil)
-        
-        switch capturedError {
-        case .serverError(let response, let data)?:
-            XCTAssertNil(response)
-            XCTAssertNil(data)
-        default:
-            XCTFail("Expect serverError")
-        }
-    }
-    
-    func testRequest_withFailingSerialization() {
-        //Given
-        networkAccess.changeMock(data: Train.JSONDataWithInvalidKey, response: nil, error: nil)
-        let expection = expectation(description: "testRequest_withFailingSerialization")
-        
-        //When
-        networkService.request(resource, onCompletion: { _ in
-            XCTFail("Should not call success block")
-        }, onError: { (error: NetworkError) in
-                if case .serializationError(_, _) = error {
-                    expection.fulfill()
-                } else {
-                    XCTFail("Expects serializationError")
-                }
-        })
-        
-        waitForExpectations(timeout: 1, handler: nil)
-    }
-    
-    func testRequest_withErrorResponse() {
-        //Given
-        let error = NSError(domain: "", code: 0, userInfo: nil)
-        networkAccess.changeMock(data: nil, response: nil, error: error)
-        let expection = expectation(description: "testOnError")
-        
-        //When
-        networkService.request(resource, onCompletion: { _ in
-            }, onError: { resultError in
-                //Then
-                switch resultError {
-                case .requestError:
-                    expection.fulfill()
-                default:
-                    XCTFail("Expects requestError")
-                }
-        })
-        
-        waitForExpectations(timeout: 1, handler: nil)
-    }
-    
-    private lazy var testData: Data! = {
-        return "test_string".data(using: .utf8)
-    }()
-    
-    func testRequest_withStatusCode401Response() {
-        //Given
-        let expectedResponse = HTTPURLResponse(url: .defaultMock, statusCode: 401, httpVersion: nil, headerFields: nil)
-        networkAccess.changeMock(data: testData, response: expectedResponse, error: nil)
-        let expection = expectation(description: "testOnError")
-        
-        //When
-        networkService.request(resource, onCompletion: { _ in
-            }, onError: { resultError in
-                //Then
-                switch resultError {
-                case .unauthorized(let response, let data):
-                    XCTAssertEqual(response, expectedResponse)
-                    XCTAssertEqual(data, self.testData)
-                    expection.fulfill()
-                default:
-                    XCTFail("Expects unauthorized")
-                }
-        })
-        
-        waitForExpectations(timeout: 1, handler: nil)
-    }
-    
-    func testGIVEN_aRequest_WHEN_requestWithResultResponse_THEN_ShouldRespond() {
-        // GIVEN
-        
-        networkAccess.changeMock(data: Train.validJSONData, response: .defaultMock, error: nil)
-        let expection = expectation(description: "loadValidRequest")
-        var expectedResult: Result<Train, NetworkError>?
-        
-        //When
-        networkService.request(resource, onCompletion: { result in
-            expectedResult = result
-            expection.fulfill()
-        })
-        
-        waitForExpectations(timeout: 1, handler: nil)
-        
-        //Then
-        switch expectedResult {
-        case .success(let train)?:
-            XCTAssertEqual(train.name, self.trainName)
-        case .failure?:
-            XCTFail("Should be an error")
-        case nil:
-            XCTFail("Result should not be nil")
-        }
-        XCTAssertEqual(networkAccess.request?.url?.absoluteString, "https://bahn.de/train")
-    }
-    
-    func testGIVEN_aRequest_WHEN_requestWithResultErrorResponse_THEN_ShouldError() {
-        //Given
-        networkAccess.changeMock(data: nil, response: nil, error: nil)
-        var expectedResult: Result<Train, NetworkError>?
-        let expection = expectation(description: "testNoData")
-        
-        //When
-        
-        networkService.request(resource, onCompletion: { result in
-            expectedResult = result
-            expection.fulfill()
-        })
-        
-        //Then
-        waitForExpectations(timeout: 1, handler: nil)
-        
-        switch expectedResult {
-        case .failure(let error)?:
-            if case .serverError(let response, let data) = error {
-                XCTAssertNil(response)
-                XCTAssertNil(data)
-            } else {
-                XCTFail("Expect serverError")
-            }
-        default:
-            XCTFail("Expect serverError")
-        }
-    }
-    
-    func testGIVEN_aRequest_WHEN_requestWithResultAndResponse_THEN_ShouldRespond() {
-        // GIVEN
-        
-        networkAccess.changeMock(data: Train.validJSONData, response: .defaultMock, error: nil)
-        let expection = expectation(description: "loadValidRequest")
-        var expectedResult: Result<(Train, HTTPURLResponse), NetworkError>?
-        
-        //When
-        networkService.request(resource: resource) { (result) in
-            expectedResult = result
-            expection.fulfill()
-        }
-        
-        waitForExpectations(timeout: 1, handler: nil)
-        
-        //Then
-        switch expectedResult {
-        case .success(let result)?:
-            XCTAssertEqual(result.0.name, self.trainName)
-            XCTAssertEqual(result.1, .defaultMock)
-        case .failure?:
-            XCTFail("Should be an error")
-        case nil:
-            XCTFail("Result should not be nil")
-        }
-        XCTAssertEqual(networkAccess.request?.url?.absoluteString, "https://bahn.de/train")
-    }
-
-    @available(macOS 10.15, iOS 13.0, watchOS 6.0, tvOS 13.0, *)
-    func testGIVEN_aRequest_WHEN_requestWithAsyncResultAndResponse_THEN_ShouldRespond() async throws {
-        // GIVEN
-        networkAccess.changeMock(data: Train.validJSONData, response: .defaultMock, error: nil)
+        let networkAccess = NetworkAccessMock(result: .success((Train.validJSONData, HTTPURLResponse.defaultMock)))
+        let networkService = BasicNetworkService(networkAccess: networkAccess)
 
         //When
-        let (result, response) = try await networkService.request(resource)
-
+        let (train, response) = try await networkService.requestResultWithResponse(for: resource).get()
 
         //Then
-        XCTAssertEqual(result.name, self.trainName)
+        XCTAssertEqual(train.name, self.trainName)
         XCTAssertEqual(response, .defaultMock)
-        XCTAssertEqual(networkAccess.request?.url?.absoluteString, "https://bahn.de/train")
+        let request = await networkAccess.request
+        XCTAssertEqual(request?.url?.absoluteString, "https://bahn.de/train")
+    }
+    
+    func testRequest_withFailingSerialization() async {
+        //Given
+        let networkAccess = NetworkAccessMock(result: .success((Train.JSONDataWithInvalidKey, HTTPURLResponse.defaultMock)))
+        let networkService = BasicNetworkService(networkAccess: networkAccess)
+        
+        //When
+        let result = await networkService.requestResult(for: resource)
+
+        //Then
+        if case .failure(.serializationError(_, let data)) = result {
+            XCTAssertEqual(data, Train.JSONDataWithInvalidKey)
+        } else {
+            XCTFail("Expects serializationError")
+        }
+    }
+    
+    func testRequest_withErrorResponse() async {
+        //Given
+        let error = NSError(domain: "", code: 0, userInfo: nil)
+        let networkAccess = NetworkAccessMock(result: .failure(error))
+        let networkService = BasicNetworkService(networkAccess: networkAccess)
+
+        //When
+        let result = await networkService.requestResult(for: resource)
+
+        //Then
+        switch result {
+        case .failure(.requestError):
+            return
+        default:
+            XCTFail("Expects requestError")
+        }
     }
 
-    @available(macOS 10.15, iOS 13.0, watchOS 6.0, tvOS 13.0, *)
-    func testGIVEN_aRequest_WHEN_requestWithAsyncResultAndResponse_THEN_ShouldThwo() async {
-        // GIVEN
+    func testRequest_withErrorThrow() async {
+        //Given
         let error = NSError(domain: "", code: 0, userInfo: nil)
-        networkAccess.changeMock(data: nil, response: nil, error: error)
+        let networkAccess = NetworkAccessMock(result: .failure(error))
+        let networkService = BasicNetworkService(networkAccess: networkAccess)
 
         //When
         do {
             try await networkService.request(resource)
-            XCTFail("Schould throw")
-        } catch let error {
-            XCTAssertTrue(error is NetworkError)
+            XCTFail("Expects throws")
+        } catch {
+            return
         }
     }
 
-    @available(macOS 10.15, iOS 13.0, watchOS 6.0, tvOS 13.0, *)
-    func testGIVEN_aRequest_WHEN_requestWithAsyncResultAndResponseAndCancel_THEN_ShouldThwo() async {
-        // GIVEN
-        let error = NSError(domain: "", code: 0, userInfo: nil)
-        networkAccess.changeMock(data: nil, response: nil, error: error)
+    func testRequest_withStatusCode401Response() async throws {
+        //Given
+        let testData: Data! = "test_string".data(using: .utf8)
+        let expectedResponse = try XCTUnwrap(HTTPURLResponse(url: .defaultMock, statusCode: 401, httpVersion: nil, headerFields: nil))
+        let networkAccess = NetworkAccessMock(result: .success((testData, expectedResponse)))
+        let networkService = BasicNetworkService(networkAccess: networkAccess)
 
         //When
-        let task = Task {
-            try await networkService.request(resource)
-        }
-        task.cancel()
-        let result = await task.result
-        if case .failure(let error) = result, let networkError = error as? CancellationError {
-           
+        let result = await networkService.requestResult(for: resource)
+
+        //Then
+        if case .failure(.unauthorized(response: expectedResponse, data: testData)) = result {
+            return
         } else {
-            XCTFail("Schould throw")
+            XCTFail()
         }
     }
+
 }
