@@ -35,163 +35,55 @@ enum CustomError: Error {
 
 class NetworkServiceWithErrorTest: XCTestCase {
     
-    var networkService: NetworkService!
-    
-    var networkAccess = NetworkAccessMock()
-    
     let trainName = "ICE"
     
-    var resource: ResourceWithError<Train, CustomError> {
+    var resource: Resource<Train, CustomError> {
         let request = URLRequest(path: "train", baseURL: .defaultMock)
-        return ResourceWithError(request: request, decoder: JSONDecoder(), mapError: { CustomError(networkError: $0) })
+        return Resource(request: request, decoder: JSONDecoder(), mapError: { CustomError(networkError: $0) })
     }
+
     
-    override func setUp() {
-        networkService = BasicNetworkService(networkAccess: networkAccess)
-    }
-    
-    func testRequest_withValidResponse() {
+    func testRequest_withValidResponse() async throws {
         //Given
-        networkAccess.changeMock(data: Train.validJSONData, response: .defaultMock, error: nil)
-        let expection = expectation(description: "loadValidRequest")
-        
-        //When
-        networkService.request(resource, onCompletionWithResponse: { train, response in
-            XCTAssertEqual(train.name, self.trainName)
-            XCTAssertEqual(response, .defaultMock)
-            expection.fulfill()
-            }, onError: { _ in
-                XCTFail("Should not call error block")
-        })
-        
-        waitForExpectations(timeout: 1, handler: nil)
-        
-        //Then
-        XCTAssertEqual(networkAccess.request?.url?.absoluteString, "https://bahn.de/train")
-    }
-
-    func testRequest_withError() {
-        //Given
-        networkAccess.changeMock(data: nil, response: nil, error: nil)
-        let expection = expectation(description: "testNoData")
-        
-        //When
-        var capturedError: CustomError?
-        networkService.request(resource, onCompletion: { _ in
-            XCTFail("Should not call success block")
-            }, onError: { error in
-                capturedError = error
-                expection.fulfill()
-        })
-        
-        //Then
-        waitForExpectations(timeout: 1, handler: nil)
-        
-        XCTAssertEqual(capturedError, .error)
-    }
-
-    
-    func testGIVEN_aRequest_WHEN_requestWithResultResponse_THEN_ShouldRespond() {
-        // GIVEN
-        
-        networkAccess.changeMock(data: Train.validJSONData, response: .defaultMock, error: nil)
-        let expection = expectation(description: "loadValidRequest")
-        var expectedResult: Result<Train, CustomError>?
-        
-        //When
-        networkService.request(resource, onCompletion: { result in
-            expectedResult = result
-            expection.fulfill()
-        })
-        
-        waitForExpectations(timeout: 1, handler: nil)
-        
-        //Then
-        switch expectedResult {
-        case .success(let train)?:
-            XCTAssertEqual(train.name, self.trainName)
-        case .failure?:
-            XCTFail("Should be an error")
-        case nil:
-            XCTFail("Result should not be nil")
-        }
-        XCTAssertEqual(networkAccess.request?.url?.absoluteString, "https://bahn.de/train")
-    }
-    
-    func testGIVEN_aRequest_WHEN_requestWithResultErrorResponse_THEN_ShouldError() {
-        //Given
-        networkAccess.changeMock(data: nil, response: nil, error: nil)
-        var expectedResult: Result<Train, CustomError>?
-        let expection = expectation(description: "testNoData")
-        
-        //When
-        
-        networkService.request(resource, onCompletion: { result in
-            expectedResult = result
-            expection.fulfill()
-        })
-        
-        //Then
-        waitForExpectations(timeout: 1, handler: nil)
-
-        XCTAssertEqual(expectedResult, .failure(.error))
-    }
-    
-    func testGIVEN_aRequest_WHEN_requestWithResultAndResponse_THEN_ShouldRespond() {
-        // GIVEN
-        
-        networkAccess.changeMock(data: Train.validJSONData, response: .defaultMock, error: nil)
-        let expection = expectation(description: "loadValidRequest")
-        var expectedResult: Result<(Train, HTTPURLResponse), CustomError>?
-        
-        //When
-        networkService.request(resource: resource) { (result) in
-            expectedResult = result
-            expection.fulfill()
-        }
-        
-        waitForExpectations(timeout: 1, handler: nil)
-        
-        //Then
-        switch expectedResult {
-        case .success(let result)?:
-            XCTAssertEqual(result.0.name, self.trainName)
-            XCTAssertEqual(result.1, .defaultMock)
-        case .failure?:
-            XCTFail("Should be an error")
-        case nil:
-            XCTFail("Result should not be nil")
-        }
-        XCTAssertEqual(networkAccess.request?.url?.absoluteString, "https://bahn.de/train")
-    }
-
-    @available(macOS 10.15, iOS 13.0, watchOS 6.0, tvOS 13.0, *)
-    func testGIVEN_aRequest_WHEN_requestWithAsyncResultAndResponse_THEN_ShouldRespond() async throws {
-        // GIVEN
-        networkAccess.changeMock(data: Train.validJSONData, response: .defaultMock, error: nil)
+        let networkAccess = NetworkAccessMock(result: .success((Train.validJSONData, HTTPURLResponse.defaultMock)))
+        let networkService = BasicNetworkService(networkAccess: networkAccess)
 
         //When
-        let (result, response) = try await networkService.request(resource)
-
+        let (train, response) = try await networkService.requestResultWithResponse(for: resource).get()
 
         //Then
-        XCTAssertEqual(result.name, self.trainName)
+        XCTAssertEqual(train.name, self.trainName)
         XCTAssertEqual(response, .defaultMock)
-        XCTAssertEqual(networkAccess.request?.url?.absoluteString, "https://bahn.de/train")
+        let request = await networkAccess.request
+        XCTAssertEqual(request?.url?.absoluteString, "https://bahn.de/train")
     }
 
-    @available(macOS 10.15, iOS 13.0, watchOS 6.0, tvOS 13.0, *)
-    func testGIVEN_aRequest_WHEN_requestWithAsyncResultAndResponse_THEN_ShouldThwo() async {
-        // GIVEN
+    func testRequest_withError() async {
         let error = NSError(domain: "", code: 0, userInfo: nil)
-        networkAccess.changeMock(data: nil, response: nil, error: error)
+        let networkAccess = NetworkAccessMock(result: .failure(error))
+        let networkService = BasicNetworkService(networkAccess: networkAccess)
+
+        //When
+        let result = await networkService.requestResult(for: resource)
+
+        //Then
+        XCTAssertEqual(result, .failure(.error))
+    }
+    
+    func testRequestThrows_withError() async {
+        let error = NSError(domain: "", code: 0, userInfo: nil)
+        let networkAccess = NetworkAccessMock(result: .failure(error))
+        let networkService = BasicNetworkService(networkAccess: networkAccess)
 
         //When
         do {
             try await networkService.request(resource)
-            XCTFail("Schould throw")
-        } catch let error {
-            XCTAssertTrue(error is CustomError)
+            XCTFail()
+        } catch let error as CustomError {
+            XCTAssertEqual(error, .error)
+        } catch {
+            XCTFail()
         }
     }
+
 }
