@@ -27,17 +27,25 @@ import DBNetworkStack
 
 class ResourceTest: XCTestCase {
     
-    func testResource() {
+    func testResource() throws {
         //Given
         let validData: Data! = "ICE".data(using: .utf8)
 
-        let resource = Resource<String?>(request: URLRequest.defaultMock, parse: { String(data: $0, encoding: .utf8) })
-        
+        let resource = Resource<String?, NetworkError>(request: URLRequest.defaultMock, parse: { String(data: $0, encoding: .utf8) })
+
         //When
-        let name = try? resource.parse(validData)
+        let name = try resource.parse(validData)
         
         //Then
-        XCTAssertEqual(name ?? nil, "ICE")
+        XCTAssertEqual(name, "ICE")
     }
-    
+
+    func testResourceWithVoidResult() throws {
+        //Given
+        let resource = Resource<Void, NetworkError>(request: URLRequest.defaultMock)
+
+        //When
+        try resource.parse(Data())
+    }
+
 }
