@@ -25,116 +25,101 @@ import XCTest
 @testable import DBNetworkStack
 
 class RetryNetworkserviceTest: XCTestCase {
-    var networkServiceMock: NetworkServiceMock!
-    var resource: Resource<Int> {
+    var resource: Resource<Int, NetworkError> {
         let request = URLRequest(path: "/train", baseURL: .defaultMock)
         return Resource(request: request, parse: { _ in return 1})
     }
     
-    override func setUp() {
-        super.setUp()
-        networkServiceMock = NetworkServiceMock()
-    }
-    
-    override func tearDown() {
-        networkServiceMock = nil
-        super.tearDown()
-    }
-    
-    func testRetryRequest_shouldRetry() throws {
+    func testRetryRequest_shouldRetry() async throws {
         //Given
         let errorCount = 2
         let numberOfRetries = 2
-        var executedRetrys = 0
-        
-        let retryService = RetryNetworkService(networkService: networkServiceMock, numberOfRetries: numberOfRetries,
-                                               idleTimeInterval: 0, shouldRetry: { _ in return true }, dispatchRetry: { _, block in
-                                                executedRetrys += 1
-                                                block()
-        })
+        let networkServiceMock = NetworkServiceMock(
+            Result<Int, NetworkError>.failure(.unknownError),
+            Result<Int, NetworkError>.failure(.unknownError),
+            Result<Int, NetworkError>.success(1)
+        )
+
+        let retryService = RetryNetworkService(
+            networkService: networkServiceMock,
+            numberOfRetries: numberOfRetries,
+            idleTimeInterval: 0,
+            shouldRetry: { _ in
+                return true
+            }
+        )
         
         //When
-        weak var task = retryService.request(resource, onCompletion: { _ in
-            XCTAssertEqual(executedRetrys, numberOfRetries)
-        }, onError: { _ in
-            XCTFail("Expects to not call error block")
-        })
-        try (0..<errorCount).forEach { _ in
-            try networkServiceMock.returnError(with: .unknownError)
+        let result = try await retryService.request(resource)
+
+        
+        //Then
+        XCTAssertEqual(result, 1)
+    }
+    
+//    func testRetryRequestWhenCanceld_shouldNotRetry() throws {
+//        //Given
+//        let networkServiceMock = NetworkServiceMock()
+//        let retryService = RetryNetworkService(
+//            networkService: networkServiceMock,
+//            numberOfRetries: 2,
+//            idleTimeInterval: 0,
+//            shouldRetry: { _ in return true }
+//        )
+//        let task = Task {
+//            await retryService.requestResult(for: resource)
+//        }
+//        task.cancel()
+//
+//        //When
+//       
+//        task?.cancel()
+//        try networkServiceMock.returnError(with: .unknownError)
+//        
+//        //Then
+//        XCTAssertNil(task)
+//    }
+    
+    func testRetryRequest_moreErrorsThenRetryAttemps() async throws {
+        //Given
+        let networkServiceMock = NetworkServiceMock(
+            Result<Int, NetworkError>.failure(.unknownError),
+            Result<Int, NetworkError>.failure(.unknownError),
+            Result<Int, NetworkError>.failure(.unknownError),
+            Result<Int, NetworkError>.failure(.unknownError)
+        )
+        let retryService = RetryNetworkService(
+            networkService: networkServiceMock,
+            numberOfRetries: 3,
+            idleTimeInterval: 0,
+            shouldRetry: { _ in return true }
+        )
+
+        //When
+        let result = await retryService.requestResult(for: resource)
+
+        
+        //Then
+        if case .failure(.unknownError) = result {
+
+        } else {
+            XCTFail()
         }
-        try networkServiceMock.returnSuccess(with: 1)
-        
-        //Then
-        XCTAssertNil(task)
     }
     
-    func testRetryRequestWhenCanceld_shouldNotRetry() throws {
+    func testRetryRequest_shouldNotRetry() async throws {
         //Given
-        let retryService = RetryNetworkService(networkService: networkServiceMock, numberOfRetries: 2,
-                                               idleTimeInterval: 0, shouldRetry: { _ in return true }, dispatchRetry: { _, block in
-                                                XCTFail("Expects to not call error block")
-                                                block()
-        })
-        weak var task = retryService.request(resource, onCompletion: { _ in
-            XCTFail("Expects to not call error block")
-        }, onError: { _ in
-            XCTFail("Expects to not call error block")
-        })
-        
+        let networkServiceMock = NetworkServiceMock(
+            Result<Int, NetworkError>.failure(.unknownError)
+        )
+        let retryService = RetryNetworkService(
+            networkService: networkServiceMock,
+            numberOfRetries: 3,
+            idleTimeInterval: 0,
+            shouldRetry: { _ in return true }
+        )
+
         //When
-       
-        task?.cancel()
-        try networkServiceMock.returnError(with: .unknownError)
-        
-        //Then
-        XCTAssertNil(task)
-    }
-    
-    func testRetryRequest_moreErrorsThenRetryAttemps() throws {
-        //Given
-        var executedRetrys = 0
-        
-        //When
-        weak var task: NetworkTask?
-        task = RetryNetworkService(networkService: networkServiceMock, numberOfRetries: 3,
-                                   idleTimeInterval: 0, shouldRetry: { _ in return true },
-                                   dispatchRetry: { _, block in
-            executedRetrys += 1
-            block()
-        }).request(resource, onCompletion: { _ in
-             XCTFail("Expects to not call completion block")
-        }, onError: { _ in
-            XCTAssertEqual(executedRetrys, 3)
-        })
-        try (0..<4).forEach { _ in
-            try networkServiceMock.returnError(with: .unknownError)
-        }
-        
-        //Then
-        XCTAssertNil(task)
-    }
-    
-    func testRetryRequest_shouldNotRetry() throws {
-        //Given
-        let shoudlRetry = false
-        var capturedError: NetworkError?
-        
-        //When
-        weak var task: NetworkTask?
-        task = RetryNetworkService(networkService: networkServiceMock, numberOfRetries: 3,
-                                   idleTimeInterval: 0, shouldRetry: { _ in return shoudlRetry },
-                                   dispatchRetry: { _, block in
-            XCTFail("Expects to not retry")
-            block()
-        }).request(resource, onCompletion: { _ in
-            XCTFail("Expects to not call completion block")
-        }, onError: { error in
-           capturedError = error
-        })
-        try networkServiceMock.returnError(with: .unknownError)
-        
-        //Then
-        XCTAssertNil(task)
-        XCTAssertNotNil(capturedError)
+        await retryService.requestResult(for: resource)
     }
 }
